feat(ai/territory): 채우기-우선 성장 + 자기포위 수정 + 영역 인구 float
- AI 성장: 건물 단독 배치(TryFillBesideRoad) 1순위 → 폐쇄구역 내부 채움(새 링 X),
  채울 자리 없을 때만 확장 1회. enclosure flood로 enclosed 안쪽만 채워 바깥
  프런티어 보존(자기포위 정지 해소). 확장 게이트는 적 영역만 회피(InEnemyTerritory)로
  완화 — 자기 영역 위 확장이 막혀 정지하던 버그 수정.
- Territory: PopPerCell int→float, 셀 수 = floor(인구/PopPerCell)(무조건 내림).
  TerritoryConfig 싱글톤화 + Test.cs PopPerCell(float) 런타임 튜닝.
- UI 영어화: PreviewStatusOps.ToText / RoadBuildController.HoverStatus /
  BuildingPlaceController.StatusText.
- (Unity) 거주/창고/화약 프리팹 BuildingAuthoring + 건설탭 버튼 씬 와이어링.

+++ Assets/_game/scripts/RunTime/Components/TerritoryComponents.cs
@@ -26,14 +26,17 @@ namespace CitySim
     /// </summary>
     public struct TerritoryConfig : IComponentData
     {
-        /// <summary>셀당 인구 기준. 영역 셀 수 = 인구 / PopPerCell.</summary>
-        public int PopPerCell;
+        /// <summary>
+        /// 셀 1칸 점유에 필요한 인구(충족값, float). 영역 셀 수 = floor(인구 / PopPerCell).
+        /// 인구(Capacity)는 int지만 나눗셈은 float, 나머지는 올림 없이 무조건 내림.
+        /// </summary>
+        public float PopPerCell;
         /// <summary>영역 확산 윈도우 최대 반경(셀, 성능·폭주 가드).</summary>
         public int MaxRadius;
 
         public static TerritoryConfig Default => new TerritoryConfig
         {
-            PopPerCell = 5,
+            PopPerCell = 5f,
             MaxRadius  = 64,
         };
     }
